extract server object to its own package
