+++ internal/server/middleware.go
@@ -1,15 +1,16 @@
-package main
+package server
 
 import (
 	"fmt"
-	"github.com/DataDog/datadog-go/statsd"
-	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/ext"
-	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/tracer"
 	"net/http"
 	"reflect"
 	"runtime"
 	"strings"
 	"time"
+
+	"github.com/DataDog/datadog-go/statsd"
+	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/ext"
+	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/tracer"
 )
 
 type loggedResponseData struct {
